quantum-miner/requirements.txt
redis==5.0.1
pydantic==2.6.1
requests==2.31.0
python-dotenv==1.0.1
qiskit==1.0.0
qiskit-ibm-runtime==0.20.0